Conflicts:
	create/src/schema.tmpl
	include/zbxcacheconfig.h
	include/zbxdbschema.h
	src/libs/zbxcacheconfig/dbconfig.c
	src/libs/zbxcacheconfig/dbsync.c
	src/libs/zbxdbupgrade/dbupgrade_7030.c
	src/zabbix_server/proxyconfigread/proxyconfigread.c
	ui/app/controllers/CControllerMiscConfigEdit.php
	ui/app/controllers/CControllerMiscConfigUpdate.php
	ui/app/views/administration.miscconfig.edit.php
	ui/include/classes/api/services/CSettings.php
	ui/include/classes/helpers/CSettingsHelper.php
	ui/include/defines.inc.php
	ui/include/schema.inc.php

+++ src/libs/zbxdbwrap/host.c
@@ -1321,7 +1321,7 @@ static void	DBdelete_graphs_by_itemids(const zbx_vector_uint64_t *itemids, int a
 		goto out;
 
 	sql = (char *)zbx_malloc(sql, sql_alloc);
-	zbx_hashset_create(&graph_index, itemids->values_num, ZBX_DEFAULT_UINT64_HASH_FUNC,
+	zbx_hashset_create(&graph_index, (size_t)itemids->values_num, ZBX_DEFAULT_UINT64_HASH_FUNC,
 			ZBX_DEFAULT_UINT64_COMPARE_FUNC);
 	zbx_vector_uint64_create(&graphids);
 
@@ -1347,7 +1347,7 @@ static void	DBdelete_graphs_by_itemids(const zbx_vector_uint64_t *itemids, int a
 
 	/* filter out graphs with other items */
 
-	zbx_hashset_create(&item_index, itemids->values_num, ZBX_DEFAULT_UINT64_HASH_FUNC,
+	zbx_hashset_create(&item_index, (size_t)itemids->values_num, ZBX_DEFAULT_UINT64_HASH_FUNC,
 			ZBX_DEFAULT_UINT64_COMPARE_FUNC);
 
 	for (int i = 0; i < itemids->values_num; i++)
